+++ javascript/src/local/index.ts
@@ -1,5 +1,6 @@
 export * from "./core.js";
 export * from "./context.js";
+export * from "./knowledge.js";
 export * from "./pause.js";
 export * from "./shell.js";
 export * from "./tools.js";